DepartmentService fixes Exception fixes Some minor changes

+++ src/main/java/ru/korona/task/models/AbstractWorker.java
@@ -10,5 +10,5 @@ import lombok.experimental.SuperBuilder;
 public abstract class AbstractWorker implements Worker {
     private final Integer id;
     private final String name;
-    private final Integer salary;
+    private final Double salary;
 }
